move utils methods
