+++ .secrets.baseline
@@ -3,7 +3,7 @@
     "files": "go.sum|^.secrets.baseline$",
     "lines": null
   },
-  "generated_at": "2024-05-16T07:08:24Z",
+  "generated_at": "2024-05-16T13:05:04Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -141,7 +141,16 @@
         "hashed_secret": "2ace62c1befa19e3ea37dd52be9f6d508c5163e6",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 244,
+        "line_number": 251,
+        "type": "Secret Keyword",
+        "verified_result": null
+      }
+    ],
+    "pkg/mounter/mounter-rclone_test.go": [
+      {
+        "hashed_secret": "2e7a7ee14caebf378fc32d6cf6f557f347c96773",
+        "is_verified": false,
+        "line_number": 364,
         "type": "Secret Keyword",
         "verified_result": null
       }
@@ -152,6 +161,24 @@
         "is_secret": false,
         "is_verified": false,
         "line_number": 78,
+        "type": "Secret Keyword",
+        "verified_result": null
+      }
+    ],
+    "pkg/mounter/mounter-s3fs_test.go": [
+      {
+        "hashed_secret": "2e7a7ee14caebf378fc32d6cf6f557f347c96773",
+        "is_verified": false,
+        "line_number": 154,
+        "type": "Secret Keyword",
+        "verified_result": null
+      }
+    ],
+    "pkg/mounter/mounter_test.go": [
+      {
+        "hashed_secret": "2e7a7ee14caebf378fc32d6cf6f557f347c96773",
+        "is_verified": false,
+        "line_number": 31,
         "type": "Secret Keyword",
         "verified_result": null
       }
@@ -194,23 +221,14 @@
       {
         "hashed_secret": "c7c6508b19455e3e8040e60e9833fbede92e5d8e",
         "is_verified": false,
-        "line_number": 124,
+        "line_number": 198,
         "type": "Secret Keyword",
         "verified_result": null
       },
       {
         "hashed_secret": "788f47a335cc4d15e0cdbe8850580c6468303558",
         "is_verified": false,
-        "line_number": 171,
-        "type": "Secret Keyword",
-        "verified_result": null
-      }
-    ],
-    "tests/e2e/rclone_test.go": [
-      {
-        "hashed_secret": "d5cd5ec53cf11ad72f24451e6c077d94d0481046",
-        "is_verified": false,
-        "line_number": 63,
+        "line_number": 245,
         "type": "Secret Keyword",
         "verified_result": null
       }
@@ -219,7 +237,7 @@
       {
         "hashed_secret": "d5cd5ec53cf11ad72f24451e6c077d94d0481046",
         "is_verified": false,
-        "line_number": 73,
+        "line_number": 142,
         "type": "Secret Keyword",
         "verified_result": null
       }
